feat: fit random forests with per-row sample weights
Node statistics carry a total row weight instead of a bootstrap row count.
Without weights every term is an integer and an integer sum accumulated in
f64 is exact, so every unweighted fit is bit-for-bit unchanged: the frozen
artifact fingerprints and reference fixtures are untouched.

A weight multiplies the bootstrap replication count, and the product is what
every impurity, split threshold test, leaf mean, and leaf distribution
accumulates. The minimum split and leaf sizes therefore count weight rather
than rows, which is what makes an integer weight the same fit as repeating
the row. A weight of zero removes the row entirely, including from the
bootstrap draw, so a resample can never come back with a zero total weight.

+++ src/ensemble/random_forest/tree.rs
@@ -347,7 +347,7 @@ impl ClassTree {
     pub(super) fn from_build_nodes(
         build: Vec<BuildNode>,
         class_weights: &[f64],
-        weights: &[u64],
+        weights: &[f64],
         classes: usize,
         n_features: usize,
     ) -> Result<Self, ModelError> {
@@ -355,7 +355,7 @@ impl ClassTree {
         debug_assert_eq!(weights.len(), build.len());
         let mut probabilities = Vec::new();
         let mut push_leaf = |index: usize| {
-            let total = weights[index] as f64;
+            let total = weights[index];
             for class in 0..classes {
                 probabilities.push((class_weights[index * classes + class] / total) as f32);
             }
